@@ -5,7 +5,7 @@ description: JavaScript parser originally written by Asger Feldthaus<asger.feldt
 homepage: https://github.com/anuragvohraec/parsejs.dart.git
 
 environment:
-  sdk: '>=2.4.0 < 3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   unicode: '>=0.0.8 <1.0.0'
