@@ -8,7 +8,7 @@ void annotateAST(Program ast) {
   new Resolver()..resolve(ast);
 }
 
-void setParentPointers(Node node, [Node parent]) {
+void setParentPointers(Node node, [Node? parent]) {
   node.parent = parent;
   node.forEach((child) => setParentPointers(child, node));
 }
@@ -19,25 +19,25 @@ class EnvironmentBuilder extends RecursiveVisitor<Null> {
     visit(ast);
   }
 
-  Scope currentScope; // Program or FunctionExpression
+  Scope? currentScope; // Program or FunctionExpression
 
   void addVar(Name name) {
-    currentScope.environment.add(name.value);
+    currentScope!.environment!.add(name.value);
   }
 
   visitProgram(Program node) {
-    node.environment = new Set<String>();
+    node.environment = new Set<String?>();
     currentScope = node;
     node.forEach(visit);
   }
 
   visitFunctionNode(FunctionNode node) {
-    node.environment = new Set<String>();
-    Scope oldScope = currentScope;
+    node.environment = new Set<String?>();
+    Scope? oldScope = currentScope;
     currentScope = node;
-    node.environment.add('arguments');
+    node.environment!.add('arguments');
     if (node.isExpression && node.name != null) {
-      addVar(node.name);
+      addVar(node.name!);
     }
     node.params.forEach(addVar);
     visit(node.body);
@@ -45,7 +45,7 @@ class EnvironmentBuilder extends RecursiveVisitor<Null> {
   }
 
   visitFunctionDeclaration(FunctionDeclaration node) {
-    addVar(node.function.name);
+    addVar(node.function.name!);
     visit(node.function);
   }
 
@@ -59,8 +59,8 @@ class EnvironmentBuilder extends RecursiveVisitor<Null> {
   }
 
   visitCatchClause(CatchClause node) {
-    node.environment = new Set<String>();
-    node.environment.add(node.param.value);
+    node.environment = new Set<String?>();
+    node.environment!.add(node.param.value);
     node.forEach(visit);
   }
 }
@@ -71,17 +71,17 @@ class Resolver extends RecursiveVisitor<Null> {
     visit(ast);
   }
 
-  Scope enclosingScope(Node node) {
+  Scope enclosingScope(Node? node) {
     while (node is! Scope) {
-      node = node.parent;
+      node = node!.parent;
     }
-    return node as Scope;
+    return node;
   }
 
   Scope findScope(Name nameNode) {
-    String name = nameNode.value;
-    Node parent = nameNode.parent;
-    Node node = nameNode;
+    String? name = nameNode.value;
+    Node? parent = nameNode.parent;
+    Node? node = nameNode;
     if (parent is FunctionNode && parent.name == node && !parent.isExpression) {
       node = parent.parent;
     }
@@ -89,7 +89,7 @@ class Resolver extends RecursiveVisitor<Null> {
     while (scope is! Program) {
       if (scope.environment == null)
         throw "$scope does not have an environment";
-      if (scope.environment.contains(name)) return scope;
+      if (scope.environment!.contains(name)) return scope;
       scope = enclosingScope(scope.parent);
     }
     return scope;
