@@ -4,7 +4,7 @@ import 'package:jsparser/src/ast.dart';
 
 /// Converts an AST to a JSON object matching the AST structure emitted by the Esprima parser.
 /// This is for testing purposes, so the output can be compared against another well-tested parser.
-class Ast2Json extends Visitor<Object> {
+class Ast2Json extends Visitor<Object?> {
   bool ranges;
   bool lines;
 
@@ -13,16 +13,16 @@ class Ast2Json extends Visitor<Object> {
   // NOTE: The order in which properties are mentioned is significant, since properties
   //       must be mentioned in the same order for our JSON comparator to work.
 
-  List<Object> list(List<Node> nodes) => nodes.map(visit).toList();
+  List<Object?> list(List<Node?> nodes) => nodes.map(visit).toList();
 
-  visit(Node node) {
+  visit(Node? node) {
     if (node == null) return null;
-    Map json = node.visitBy(this) as Map;
+    Map? json = node.visitBy(this) as Map?;
     if (ranges) {
-      json['range'] = [node.start, node.end];
+      json!['range'] = [node.start, node.end];
     }
     if (lines) {
-      json['line'] = node.line;
+      json!['line'] = node.line;
     }
     return json;
   }
@@ -244,10 +244,10 @@ class Ast2Json extends Visitor<Object> {
   visitNameExpression(NameExpression node) => visit(node.name);
 
   // Some values cannot be encoded in JSON. We simply represent these as null.
-  bool isUnencodable(Object x) =>
+  bool isUnencodable(Object? x) =>
       x == double.infinity || x == double.negativeInfinity || x == double.nan;
 
-  visitLiteral(LiteralExpression node) => <String, Object>{
+  visitLiteral(LiteralExpression node) => <String, Object?>{
         'type': 'Literal',
         'value': isUnencodable(node.value) ? null : node.value,
         'raw': node.raw
